requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2
django-allauth==0.61.1
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-json-widget==1.1.1
django-jsonfield-backport==1.0.5
django-phonenumber-field==7.3.0
django-prettyjson==0.4.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.2
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1
drf-yasg==1.21.7
frozenlist==1.4.1
future==1.0.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsoneditor==1.6.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
magic-filter==1.0.12
Markdown==3.5.2
multidict==6.0.5
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.31
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyperclip==1.8.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
standardjson==0.3.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
